// web/platform-admin-web/src/api/generated/model/getTermDocumentItemStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 약관 상태
 */
export type GetTermDocumentItemStatus = typeof GetTermDocumentItemStatus[keyof typeof GetTermDocumentItemStatus];


export const GetTermDocumentItemStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  TERMINATED: 'TERMINATED',
} as const;