// web/platform-admin-web/src/api/generated/model/getTermDocumentVersionItemStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 버전 상태
 */
export type GetTermDocumentVersionItemStatus = typeof GetTermDocumentVersionItemStatus[keyof typeof GetTermDocumentVersionItemStatus];


export const GetTermDocumentVersionItemStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
} as const;